pyproject.toml
[project]
name = "jarvis-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "langgraph>=0.4.8,<0.5.0",
    "langgraph-checkpoint-sqlite>=0.1.0",
    "llama-index-core>=0.14.4",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-vector-stores-postgres>=0.7.0",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.9",
    "sentence-transformers>=5.1.1",
    "langchain-core>=0.3.14",
    "requests-toolbelt>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "typing-inspect>=0.9.0",
    "annotated-types>=0.1",
    "packaging>=20.0",
    "requests>=2.0.0",
    "urllib3>=1.26",
    "idna>=2.0",
    "certifi>=2020.0.0",
    "charset-normalizer>=3.0.0",
    "google-generativeai>=0.5.0",
    "pydantic_core",
    "typing_inspection",
    "langsmith"
]

[tool.setuptools]
packages = ["graph", "memory", "perception", "reasoning", "tools", "utils"]
